Fixed bug in passing integers to idInterpreter::EnterFunctionVarArg function.
If something is passed to vararg function as integer, it must be extracted from stack as integer later.
Hence the need to use 'd' mode instead of 'f' (when parameter has type double).
Note that the extracted integer in this case must be converted to float before pushing onto script stack, since scripts have no integers.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@7101 49c82d7f-2e2a-0410-a16f-ae8f201b507f