Fixing the mess with SIMD changes.
1. Fixed AVX detection.
It is easy to google "AVX detection", review a few links, and understand how it should work.
Why needing to reinvent a bad wheel? Windows version is not the thing to rely on.

2. Removed hacks in idPlane::Distance and with sse4_present.
You should not commit stuff like this.
And anyway, DP is horribly slow --- scalar code is surely faster, given that it avoids stalls.

3. Renamed SSE 4 to SSE 4.1.
Such things cannot be approximate.

4. Fixed print mask in idSimd.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@7518 49c82d7f-2e2a-0410-a16f-ae8f201b507f