GLSLProgram: disabled warnings for absent uniforms, do not rely on currentProgram when doing glUseProgram.
The warnings will produce a lot of spam for material-new-stage shaders, where you simply cannot determine which uniforms must be used and which are not.
Also, they are not convenient in general when you want to use a pack of uniforms (like all sort of GL matrices) easily available in several programs.

The currentProgram check is perfectly good, but only when everyone is using GLSLProgram for shaders.
Right now we have tons of code calling qglUseProgram directly, in which case this check makes it impossible to set program sometimes.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@8045 49c82d7f-2e2a-0410-a16f-ae8f201b507f