tdm_installer: check that TDM packages can be written to before installing.
This is done at the end of Actions::ScanInstallDirectoryIfNecessary.
We iterate over all TDM-owned zips and the manifest, and check that all of them is "writable".

Writable checks is basically trying to open the file in append mode.
On Windows we actually open files in full-exclusive mode, because ordinary writes to pk4 files are allowed when TDM reads from them.
On Linux this check does not help, because TDM does not lock pk4 files at all: everyone can overwrite them without any problem =(


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@9140 49c82d7f-2e2a-0410-a16f-ae8f201b507f