tdm_installer. Forbid closing installer window if some action is in progress.
Previously the window could be closed at any moment, but the process continued to work until it finished the current action.
It was quite counter-intuitive. Now close button simply does not work if some action is in progress, which is much better.

Note that we only support interruption of downloading the data during installation (usign Cancel button).
All the other actions are non-interruptible, although user can always just kill process.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@9397 49c82d7f-2e2a-0410-a16f-ae8f201b507f