#5599. Supporting lifting grabbed AF into air with new drag implementation.
Method CForce_Grab::SetFrictionOverride temporarily resets friction coefficients on the whole dragged body (only used on AF now).
The original friction coefficients are saved in CForce_Grab::m_originalFriction, saved/restored in savegame file.

It is important to revert this temporary override when object is no longer dragged.
This is ensured by:
  m_drag.SetPhysics( NULL, ... ) in CGrabber::StopDrag
  SetFrictionOverride( false ) in CForce_Grab::SetPhysics

The main changes are in CForce_Grab::Evaluate:
  1) Choose cv_drag_af_weight_ratio_canlift / cv_drag_af_weight_ratio depending on AF's "ground when dragged" flag.
  2) Detect if AF is fully in-air.
  3) If AF is fully in-air, apply force to all of its bodies uniformly instead of to one selected body only.
  4) If AF is fully in-air, then temporarily override its friction coefficients with cv_drag_af_inair_friction to avoid oscillations.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@9686 49c82d7f-2e2a-0410-a16f-ae8f201b507f