Volumetric light refactoring, major changes in volumetric fog implementation.
1) Force disable volumetric shadows for fog light.
2) Set projection/falloff textures to _white, so that material textures don't affect the result (just like with normal fog lights).
3) Use translucency-like blending equation for volumetric fog: it should partly overwrite background color, not just highlight/add to it.
4) Use exponential alpha generation mode for volumetric fog, not just something linear and unlimited (as in volumetric lights).

The alpha generation code for volumetric code is what took the most work.
The fog is exponential at its heart:
  background_coeff = exp(-5.0 * distThroughFog / distCap)
The distCap is taken from shaderParm3 as usual, but if it is not set, then it is taken from 1 / volumetric_dust.
It seems that this equation almost matches the attenuation of an ordinary fog light.

One issue is that fog should become fully opaque at distThroughFog >= distCap.
To achieve that, I hacked coefficient to snap it to zero at distCap (the hack starts at 90% of distCap to make it hardly visible).

Finally, I added a bit of dithering under the save r_volumetricDither cvar.
Without it, there is some color banding lines briefly visible at almost invisible distances, even at 64-bit colors.
The implementation and strength of dithering is not mathematically correct here, I just added some amount that felt right to me.

See also:
  https://forums.thedarkmod.com/index.php?/topic/21241-how-to-create-a-foggy-cellar-aka-make-use-of-lights-particles/&do=findComment&comment=468810


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@9804 49c82d7f-2e2a-0410-a16f-ae8f201b507f